+++ routes/notifications.js
@@ -3,6 +3,10 @@ const db = require("../database/connection");
 const router = express.Router();
 
 router.post('/', (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     const { user_id, message, createdBy } = req.body;
     const query = 'INSERT INTO notifications (user_id, message, created_by) VALUES (?, ?, ?)';
     db.query(query, [user_id, message, createdBy], async (err, result) => {
@@ -12,6 +16,10 @@ router.post('/', (req, res) => {
 });
 
 router.get('/:user_id', (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     const { user_id } = req.params;
     const query = 'SELECT * FROM notifications WHERE user_id = ? ORDER BY created_at DESC';
     db.query(query, [user_id], (err, results) => {
@@ -21,6 +29,10 @@ router.get('/:user_id', (req, res) => {
 });
 
 router.put('/:id', (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     const { id } = req.params;
     const query = 'UPDATE notifications SET is_read = TRUE WHERE id = ?';
     db.query(query, [id], (err, result) => {
@@ -30,6 +42,10 @@ router.put('/:id', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     const { id } = req.params;
     const query = 'DELETE FROM notifications WHERE id = ?';
