+++ routes/rent.js
@@ -4,6 +4,10 @@ const router = express.Router();
 
 
 router.post("/", (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     const {
         user_id, city, model, price, trim, kilometers, year, manufacturer, seats,
         car_plate_number, warranty, steering_wheel, seller_type, body,
@@ -39,6 +43,10 @@ router.post("/", (req, res) => {
 });
 
 router.get("/", (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     db.query("SELECT * FROM rent", (err, results) => {
         if (err) return res.status(500).json({ error: err.message });
         res.json(results);
@@ -47,6 +55,10 @@ router.get("/", (req, res) => {
 
 
 router.put("/:id", (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     const {
         user_id, city, model, price, trim, kilometers, year, manufacturer, seats,
         car_plate_number, warranty, steering_wheel, seller_type, body,
@@ -76,6 +88,10 @@ router.put("/:id", (req, res) => {
 });
 
 router.delete("/:id", (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     db.query("DELETE FROM rent WHERE id = ?", [req.params.id], (err, result) => {
         if (err) return res.status(500).json({ error: err.message });
         if (result.affectedRows === 0) return res.status(404).json({ message: "Ad not found" });
@@ -84,6 +100,10 @@ router.delete("/:id", (req, res) => {
 });
 
 router.get("/:id", (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     db.query("SELECT * FROM rent WHERE id = ?", [req.params.id], (err, result) => {
         if (err) return res.status(500).json({ error: err.message });
         if (result.length === 0) return res.status(404).json({ message: "Ad not found" });
@@ -92,6 +112,10 @@ router.get("/:id", (req, res) => {
 });
 
 router.get("/user/:id", (req, res) => {
+    const token =  req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ loggedIn: false, message: 'No token found' });
+    }
     db.query("SELECT * FROM rent WHERE user_id = ?", [req.params.id], (err, result) => {
         if (err) return res.status(500).json({ error: err.message });
         const formattedData = result.map(ad => ({
